rbh-update-path: save the current parent
The current parent for each batch of children is kept in memory. We
now retrieve the parent only once per batch.

We only need to retrieve the parent for each entry in the first batch
(from the mirror), because we cannot be sure that all entries share
the same parent at that stage.

Change-Id: I9dc21ec4f85e794616be13a3ac3cc2bc4d2b42d6
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>
Reviewed-on: https://gerrithub.io/c/robinhood-suite/robinhood4/+/1227952
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>